requirements.txt
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
librosa==0.10.1
tqdm==4.66.5
pyyaml==6.0.2
pandas==2.0.3
einops==0.8.0
pyworld==0.3.4